Prueba de bodega

+++ modulos/db/prueba.py
@@ -1,27 +1,27 @@
 import json
+from os import path
 
-data = {}
-
-data['clients'] = []
-
-data['clients'].append({
-    'first_name': 'Sigrid',
-    'last_name': 'Mannock',
-    'age': 27,
-    'amount': 7.17})
-
-data['clients'].append({
-    'first_name': 'Joe',
-    'last_name': 'Hinners',
-    'age': 31,
-    'amount': [1.90, 5.50]})
-
-data['clients'].append({
-    'first_name': 'Theodoric',
-    'last_name': 'Rivers',
-    'age': 36,
-    'amount': 1.11})
-
-with open('data.json', 'w') as file:
-    json.dump(data, file, indent=4)
-    
+def guardardatos():
+    data = []
+    if path.exists('data.json'):
+        with open('data.json','r') as file:
+            flores = json.load(file)
+            for flor in flores:
+                data.append(flor)
+        data.append({'nombre':"g",'tamano':'L','cantidad': 100})
+        print(data)
+        with open('data.json','w') as file:
+            json.dump(data,file, indent=4)
+    else:
+        data = []
+        Flores = {}
+        Flores['nombre'] = 'r'
+        Flores['tamano'] = 'S'
+        Flores['cantidad'] = 50
+        data.append(Flores)
+        with open('data.json','w') as file:
+            json.dump(data, file, indent=4)
+              
+guardardatos()
+  
+
